refactor : 스케줄 도착 결과 dto 패키지 분류 및 변수명 리팩토링
